Feature/account

+++ finance-tracker-backend/src/main/java/com/finance/tracker/accounts/domain/ATSnapshotCreateEvent.java
@@ -3,17 +3,19 @@ package com.finance.tracker.accounts.domain;
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
 
+import java.math.BigDecimal;
+
 @Getter
 public class ATSnapshotCreateEvent extends ApplicationEvent {
 
     private final String accountId;
     private final String transactionId;
-    private final double previousBalance;
-    private final double newBalance;
-    private final double transactionAmount;
+    private final BigDecimal previousBalance;
+    private final BigDecimal newBalance;
+    private final BigDecimal transactionAmount;
 
     public ATSnapshotCreateEvent(Object source, String accountId, String transactionId,
-                                 double previousBalance, double newBalance, double transactionAmount) {
+                                 BigDecimal previousBalance, BigDecimal newBalance, BigDecimal transactionAmount) {
         super(source);
         this.accountId=accountId;
         this.transactionId=transactionId;
