+++ finance-tracker-backend/src/main/java/com/finance/tracker/accounts/domain/NetworthSummary.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,14 +15,14 @@ public class NetworthSummary {
 
     private ValueNumber assets;
     private ValueNumber liabilities;
-    private double netWorth;
+    private BigDecimal netWorth;
 
     @Data
     @AllArgsConstructor
     @NoArgsConstructor
     public static class ValueNumber {
 
-        private double total;
+        private BigDecimal total;
         private int number;
 
     }
