+++ finance-tracker-backend/src/main/java/com/finance/tracker/accounts/domain/entities/AccountTransactionSnapshot.java
@@ -24,7 +24,7 @@ public class AccountTransactionSnapshot {
     private String transactionId;
     private BigDecimal balanceBefore;
     private BigDecimal balanceAfter;
-    private Double transactionAmount;
+    private BigDecimal transactionAmount;
     private LocalDateTime createdAt;
 
 }
